move request contracts to contracts assembly
