+++ TimesynqServer.Contracts/RequestDTO/Follow/FollowRequestDTO.cs
@@ -1,9 +1,9 @@
-﻿namespace TimesynqServer.DTO.Request.Follow
+﻿namespace TimesynqServer.Contracts.RequestDTO.Follow
 {
     /// <summary>
-    /// Represents the required information for an unfollow request.
+    /// Represents the required information for a follow request.
     /// </summary>
-    public sealed class UnfollowRequestDTO
+    public sealed class FollowRequestDTO
     {
         /// <summary>
         /// The ID of the user who is being followed.
