+++ TimesynqServer.Contracts/RequestDTO/Follow/UnfollowRequestDTO.cs
@@ -1,9 +1,9 @@
-﻿namespace TimesynqServer.DTO.Request.Follow
+﻿namespace TimesynqServer.Contracts.RequestDTO.Follow
 {
     /// <summary>
-    /// Represents the required information for a follow request.
+    /// Represents the required information for an unfollow request.
     /// </summary>
-    public sealed class FollowRequestDTO
+    public sealed class UnfollowRequestDTO
     {
         /// <summary>
         /// The ID of the user who is being followed.
