+++ TimesynqServer/TimesynqServer.csproj
@@ -17,6 +17,7 @@
   <ItemGroup>
     <ProjectReference Include="..\TimesynqServer.Application\TimesynqServer.Application.csproj" />
     <ProjectReference Include="..\TimesynqServer.Common\TimesynqServer.Common.csproj" />
+    <ProjectReference Include="..\TimesynqServer.Contracts\TimesynqServer.Contracts.csproj" />
     <ProjectReference Include="..\TimesynqServer.Infrastructure\TimesynqServer.Infrastructure.csproj" />
     <ProjectReference Include="..\TimesynqServer.Persistence\TimesynqServer.Persistence.csproj" />
   </ItemGroup>
